Split the URLPattern implementation into multiple files
The current implementation is already rather large at ~2.5k lines (~2k
LoC). There is still quite a lot of functionality left to implement,
so let's split it up while it's still manageable.
